refactor(compositor): compose_frame's geometry becomes plan_frame
The first 353 lines of `compose_frame` contained no GPU call at all — the first
one is `self.begin(...)`. They decide WHERE every layer goes: preset placements,
the app-resolved screen and camera rects, user crop, zoom regions, Full Camera,
cover/contain fitting, corner radii in render-target px, the reactive camera
shrink, and the previous-frame rects the motion blur derives velocity from.

They now live in `frame_geometry::plan_frame`, which both backends can call. The
Metal engine cannot drive a single layer today because this arithmetic was locked
inside the D3D11 engine; this is what unlocks it.

Of the 75 locals that block produces, **15 survive to the first draw**. That is the
whole interface: `scene_preset`, `mb_taps`, `source_t`, `zoom_rotation`,
`padding_scale`, `cut`, `s_dst`, `s_dst_prev`, `s_radius`, `frame_min_px`, `w_dst`,
`w_dst_prev`, `w_px`, `w_radius`, `shape_fade`. The other 60 die where they were
computed, which is why a struct of outputs is the honest shape rather than a draw
list — every `LayerCB` in this function is still built at its own draw site.

**The draw half does not change. At all.** Verified mechanically against `HEAD`, not
asserted: `self.begin(...)` to the closing brace is byte-identical. The moved block
is likewise the original text, with exactly four substitutions, each mechanical:

    self.rw() / self.rh()             -> rw / rh          (22 sites)
    scene_ref.as_ref()                -> scene            (13)
    cursor_ref.as_ref()               -> cursor            (1)
    self.timeline_t_override.borrow() -> input.timeline_t_override

The three `RefCell` guards stay in `compose_frame` and are borrowed across the call:
`scene_ref` because the draw half still reads it (`:1686`, `:1894`), `lp` because it
does too (10 sites), `cursor_ref` because the draw half rebinds its own. `plan_frame`
receives `Option<&Scene>` / `Option<&CursorTrack>` through them, so no borrow lives
one instruction longer than it did.

Nothing in the input struct is a backend object — dimensions, the scene, the live
params. `screen_tex_px` is called out in its doc because it is the one place the two
platforms genuinely differ: D3D11VA pads to macroblocks (1080 -> 1088), CoreVideo
does not, and `u_max`/`v_max` exist precisely so the geometry never assumes
texture == visible.

93 tests still pass on macOS; the Windows job type-checks the call site.
