+++ crates/compositor/src/compositor_windows.rs
@@ -1092,358 +1092,30 @@ impl Compositor {
         let u_max = scw / stw as f32;
         let v_max = sch / sth as f32;
 
-        // Scène de l'app présente → placements du layout preset (ou, mieux, le rect résolu par
-        // l'app dans `layout.webcam_rect`) ; sinon planning fixture (bench).
         let scene_ref = self.scene.borrow();
-        let scene_preset: Option<String> =
-            scene_ref.as_ref().map(|s| s.layout.preset.clone());
-        // Webcam rect résolu par l'app (= `computeCompositeLayout`, source de vérité unique
-        // entre preview et natif) : quand il est présent ET que la scène est posée, on l'utilise
-        // COMME placement de base. Sinon, fallback sur `preset_placements` historique (PiP
-        // codé en dur à 320 px + 40 px de marge — l'arrangement qui dérivait de la preview).
-        let app_webcam_rect: Option<[f32; 4]> = scene_ref
-            .as_ref()
-            .and_then(|s| s.layout.webcam_rect)
-            .map(|r| [r.x, r.y, r.width, r.height]);
-        // Idem pour l'écran. Les deux rects viennent du MÊME appel `computeCompositeLayout`, donc
-        // les consommer ensemble est la seule façon de garder le bloc écran+caméra cohérent :
-        // n'en prendre qu'un revenait à mélanger la géométrie de l'app et un placement fixture.
-        let app_screen_rect: Option<[f32; 4]> = scene_ref
-            .as_ref()
-            .and_then(|s| s.layout.screen_rect)
-            .map(|r| [r.x, r.y, r.width, r.height]);
-        let (mut p, mut pp) = match &scene_preset {
-            Some(preset) => {
-                // Chaque rect résolu par l'app remplace INDÉPENDAMMENT sa contrepartie du
-                // preset ; sinon celle du preset reste (le padding slider l'insèrera ensuite
-                // dans `scale_frame`).
-                //
-                // Avant, ce match portait sur `app_webcam_rect` et le rect ÉCRAN n'était donc
-                // honoré que si un rect webcam arrivait aussi. Un layout sans caméra gardait
-                // l'écran plein cadre du preset — pendant que `fit_screen` (plus bas) coupait
-                // quand même son fit au ratio du crop, puisqu'un `app_screen_rect` était bien
-                // présent. Résultat : un clip recadré sans caméra était étiré, et aucune des
-                // deux voies ne le rattrapait. Coupler l'écran à la présence de la caméra
-                // n'avait aucune raison d'être — ce sont deux calques indépendants.
-                let mut fp = preset_placements(preset);
-                if let Some(wr) = app_webcam_rect {
-                    fp.webcam.dst = wr;
-                }
-                if let Some(sr) = app_screen_rect {
-                    fp.screen.dst = sr;
-                }
-                (fp, fp) // layout statique → vélocité nulle
-            }
-            None => (timeline(frame, cfg), timeline(frame - 1.0, cfg)),
-        };
-        let lp = *self.live_params.borrow();
-        // Motion blur écran : quand la scène (contrat de l'app) est posée, c'est elle qui pilote
-        // (parité inspector : 1.0 + motion_blur*15 taps), sinon on retombe sur `cfg.mblur_n`
-        // (le bench fixture continue d'utiliser ses taps explicites).
-        let mb_taps = scene_ref
-            .as_ref()
-            .map(|s| 1.0 + s.effects.motion_blur.clamp(0.0, 1.0) * 15.0)
-            .unwrap_or(cfg.mblur_n as f32);
-
-        // Zoom regions + Full Camera : filtrées en amont pour le clip actif et échantillonnées
-        // dans le même référentiel source que le PTS du décodeur écran.
-        let empty_zoom: Vec<crate::scene::SceneZoomRegion> = Vec::new();
-        let empty_cam: Vec<crate::scene::SceneCameraFullscreenRegion> = Vec::new();
-        let zoom_regions = scene_ref.as_ref().map(|s| &s.zoom_regions).unwrap_or(&empty_zoom);
-        let cam_regions =
-            scene_ref.as_ref().map(|s| &s.camera_fullscreen_regions).unwrap_or(&empty_cam);
-        let webcam_reactive = scene_ref.as_ref().map(|s| s.layout.webcam_reactive_zoom).unwrap_or(false);
-        let source_t = self.timeline_t_override.borrow().unwrap_or(frame / FPS);
-        let source_t_prev = source_t - 1.0 / FPS;
-        // le focus "auto" (suivi curseur) réutilise la même piste que le rendu du curseur.
         let cursor_ref = self.cursor.borrow();
-        let cursor_for_zoom = cursor_ref.as_ref();
-        // La rotation 3D (mode 8, pas de motion blur dans ce chemin — cf. le commentaire au
-        // point d'appel) n'est calculée QUE pour la frame courante ; `pp` ne sert qu'au zoom
-        // écran normal (vélocité pour le motion blur du chemin non-tilté).
-        let mut zoom_rotation = [0.0f32; 3];
-        if !zoom_regions.is_empty() {
-            let zs = crate::regions::zoom_state_at(zoom_regions, source_t, cursor_for_zoom);
-            p.zoom = zs.scale;
-            p.focus = zs.focus;
-            zoom_rotation = zs.rotation;
-            let zs_p = crate::regions::zoom_state_at(zoom_regions, source_t_prev, cursor_for_zoom);
-            pp.zoom = zs_p.scale;
-            pp.focus = zs_p.focus;
-        }
-        // Full Camera ignore le rétrécissement réactif de la webcam (design web : mélanger
-        // "rétrécit pour le zoom" et "grandit en plein cadre" dans la même frame n'a pas de sens).
-        let cam_progress = crate::regions::camera_fullscreen_progress_at(cam_regions, source_t);
-        let cam_progress_prev =
-            crate::regions::camera_fullscreen_progress_at(cam_regions, source_t_prev);
-        // rétrécissement réactif : la webcam rétrécit pendant un zoom actif (1/zoom, plancher
-        // 0.35 — parité `reactiveWebcamScale`, TS). Ignoré pendant Full Camera (voir ci-dessus).
-        let reactive_scale = |zoom: f32, progress: f32| -> f32 {
-            if webcam_reactive && progress <= 0.0 && zoom.is_finite() && zoom > 0.0 {
-                (1.0 / zoom).clamp(0.35, 1.0)
-            } else {
-                1.0
-            }
-        };
-        // `lp.webcam_size_scale` vient de `scene.layout.webcamSize` (voir `live_params_from_scene`)
-        // — le MÊME nombre que le fraction webcamSizePreset déjà pris en compte côté app pour
-        // calculer `wr` (`computeCompositeLayout`, TS). Quand l'app fournit un `webcam_rect`
-        // explicite, la taille y est donc déjà cuite : réappliquer `lp.webcam_size_scale` ici
-        // double-échelonnerait la boîte (ex. un preset 34% → webcam rendue à ~34%×34% ≈ 12% au
-        // lieu de 34%, la webcam apparaissant bien plus petite que ce que montre l'aperçu web).
-        // Seul `reactive_scale` (rétrécissement pendant un zoom, une valeur ANIMÉE par frame que
-        // le rect statique de l'app ne capture pas) doit encore s'appliquer dans ce cas.
-        let base_size_scale = if app_webcam_rect.is_some() { 1.0 } else { lp.webcam_size_scale };
-        let webcam_size_scale = base_size_scale * reactive_scale(p.zoom, cam_progress);
-        let webcam_size_scale_prev = base_size_scale * reactive_scale(pp.zoom, cam_progress_prev);
-
-        // padding : échelle globale du layout autour du centre du cadre (parité web frameRenderer :
-        // paddingScale = 1 - padding*0.4 → padding 0 = plein cadre). S'applique à TOUS les presets :
-        // côté web, side-by-side et top/bottom soudent écran+caméra en un bloc unique et c'est ce
-        // bloc que le padding rétrécit (cf. `compositeLayout.ts`, branche `block`). Vertical-stack
-        // en était exempté tant qu'il était full-bleed ; il ne l'est plus.
-        let padding_scale = 1.0 - lp.padding * 0.4;
-        let scale_frame = |dst: [f32; 4], s: f32| -> [f32; 4] {
-            [0.5 + (dst[0] - 0.5) * s, 0.5 + (dst[1] - 0.5) * s, dst[2] * s, dst[3] * s]
-        };
-        // webcam : ancrée à son coin bas-droite (grandit vers le haut-gauche, pas depuis le centre).
-        let scale_corner_br = |dst: [f32; 4], s: f32| -> [f32; 4] {
-            let (brx, bry) = (dst[0] + dst[2], dst[1] + dst[3]);
-            let (nw, nh) = (dst[2] * s, dst[3] * s);
-            [brx - nw, bry - nh, nw, nh]
-        };
-        // parité web (compositeLayout) : rectangle/rounded gardent le ratio natif de la webcam ;
-        // square/circle forcent un carré (side = min). Le placement de base est carré → on ajuste
-        // ici, en gardant le coin bas-droite fixe (cohérent avec le size-scale).
-        let is_square_shape = matches!(lp.webcam_shape, 1 | 2); // circle | square
-        let cam_ar = if is_square_shape { 1.0 } else { (wcw / wch).max(0.01) };
-        let fit_cam_aspect = |dst: [f32; 4]| -> [f32; 4] {
-            let s = (dst[2] * self.rw()).min(dst[3] * self.rh()); // côté carré de base (px)
-            let (pw, ph) = if cam_ar >= 1.0 { (s, s / cam_ar) } else { (s * cam_ar, s) };
-            let (nw, nh) = (pw / self.rw(), ph / self.rh());
-            let (brx, bry) = (dst[0] + dst[2], dst[1] + dst[3]);
-            [brx - nw, bry - nh, nw, nh]
-        };
-        // Variantes ancrées au CENTRE (au lieu du coin bas-droite) de `dst`, pour le cas où
-        // `dst` vient de `app_webcam_rect` : ce rect est déjà la position que l'utilisateur a
-        // choisie/déplacée (résolue côté app via `computeCompositeLayout`, même convention
-        // centre-fraction que `cx`/`cy` dans `compositeLayout.ts`) — l'ancrer au coin bas-droite
-        // comme le fait `fit_cam_aspect` (pensé pour le placement par DÉFAUT, ancré à ce coin
-        // avec une marge fixe) réancre silencieusement la webcam glissée n'importe où d'autre à
-        // ce coin, ignorant la position réelle choisie par l'utilisateur — le bug rapporté
-        // (webcam glissée au coin bas-gauche, DOM/JSON envoyé au natif confirmant une position
-        // flush, mais rendu natif visiblement décalé). Le centre est le point fixe qui a un sens
-        // pour un rect DÉJÀ positionné par l'app ; le coin bas-droite n'a de sens que pour le
-        // placement par défaut, qui grandit depuis ce coin faute de position explicite.
-        let scale_center = |dst: [f32; 4], s: f32| -> [f32; 4] {
-            let (cx, cy) = (dst[0] + dst[2] * 0.5, dst[1] + dst[3] * 0.5);
-            let (nw, nh) = (dst[2] * s, dst[3] * s);
-            [cx - nw * 0.5, cy - nh * 0.5, nw, nh]
-        };
-        // Le ratio de sortie réel (peut différer du canvas interne 16:9 fixe) et le facteur
-        // d'étirement non uniforme que `blit_resized` appliquera en fin de pipeline — nécessaires
-        // ici (avant `undistort`, plus bas) pour que le fit ci-dessous cible le ratio de boîte tel
-        // qu'il apparaîtra APRÈS cet étirement, pas tel qu'il est dans l'espace canvas pré-étirement
-        // (sinon le fit et l'undistort composent deux corrections indépendantes et sur-rétrécissent
-        // le contenu — cf. rapport utilisateur : crop 9:16 + sortie 9:16 + padding 0% laissait
-        // quand même une grosse marge, alors que le crop correspond déjà exactement au cadre).
-        // Le crop de l'utilisateur (dialogue "Edit clip") a son PROPRE ratio (ex. une bande
-        // verticale 9:16 recadrée dans une source 16:9) — le zoom appliqué ensuite (§
-        // `screen_source_rect`) le préserve (mêmes facteurs sur les deux axes), donc c'est bien
-        // le ratio du CROP qui doit dimensionner le quad de destination, pas celui (fixe, issu
-        // du preset de layout) de `p.screen.dst`. Sans ça, le rect recadré (dont le ratio propre
-        // diffère de la boîte du preset) se retrouve étiré pour remplir cette boîte — parité web
-        // cassée : `computeCompositeLayout`/`centerRectInBounds` (TS) contiennent déjà le crop
-        // dans sa boîte en respectant son ratio, le natif ne le faisait pas (rapport utilisateur).
-        let active_crop = scene_ref.as_ref().and_then(|scene| {
-            scene.crop_by_clip.get(scene.active_clip_index).copied().flatten()
+        let lp = *self.live_params.borrow();
+        // Toute la géométrie vit dans `frame_geometry::plan_frame` — 353 lignes sans un
+        // appel GPU, partagées avec le backend Metal. Ce qui suit ce destructure est
+        // inchangé, à l'octet près.
+        let crate::frame_geometry::FrameGeometry {
+            scene_preset, mb_taps, source_t, zoom_rotation, padding_scale, cut, s_dst,
+            s_dst_prev, s_radius, frame_min_px, w_dst, w_dst_prev, w_px, w_radius, shape_fade,
+        } = crate::frame_geometry::plan_frame(&crate::frame_geometry::FrameGeometryInput {
+            render_px: [self.rw(), self.rh()],
+            screen_tex_px: [stw as f32, sth as f32],
+            screen_visible_px: [scw, sch],
+            webcam_visible_px: [wcw, wch],
+            u_max,
+            v_max,
+            frame,
+            cfg,
+            live: lp,
+            scene: scene_ref.as_ref(),
+            cursor: cursor_ref.as_ref(),
+            timeline_t_override: *self.timeline_t_override.borrow(),
         });
-        let crop_aspect = match active_crop {
-            Some(c) if c.width > 0.0001 && c.height > 0.0001 => {
-                (c.width * scw) / (c.height * sch).max(0.0001)
-            }
-            _ => scw / sch.max(0.0001),
-        };
-        // Contain (parité `centerRectInBounds`) : rétrécit `dst` (centré) pour que son ratio
-        // devienne `aspect`, sans jamais dépasser sa boîte d'origine — mais la boîte de référence
-        // doit être mesurée telle qu'elle apparaîtra APRÈS l'étirement de sortie (`dst` * ratio de
-        // sortie), pas dans l'espace canvas 16:9 pré-étirement : sinon le fit cible le mauvais
-        // ratio de boîte dès que la sortie n'est pas 16:9. `undistort` (plus bas) annule ensuite
-        // exactement ce même facteur, donc convertir le résultat en fraction canvas se fait par
-        // `/ uniform_stretch` (propriété de `undistort` : le ratio final ne dépend que de la
-        // taille de `dst` en PIXELS CANVAS, jamais du ratio de sortie choisi).
-        let fit_dst_to_aspect = |dst: [f32; 4], aspect: f32| -> [f32; 4] {
-            let box_w_px = dst[2] * self.rw();
-            let box_h_px = dst[3] * self.rh();
-            let box_ar = box_w_px / box_h_px.max(0.0001);
-            let (nw_px, nh_px) = if aspect > box_ar {
-                (box_w_px, box_w_px / aspect.max(0.0001))
-            } else {
-                (box_h_px * aspect, box_h_px)
-            };
-            let (nw, nh) = (nw_px / self.rw(), nh_px / self.rh());
-            let (cx, cy) = (dst[0] + dst[2] * 0.5, dst[1] + dst[3] * 0.5);
-            [cx - nw * 0.5, cy - nh * 0.5, nw, nh]
-        };
-        // Quand l'app a résolu la boîte écran, elle a DÉJÀ appliqué le padding (le rect est
-        // calculé contre `maxContentSize`) et l'a DÉJÀ mise au ratio du crop
-        // (`computeCompositeLayout` reçoit la taille de la source recadrée) : rejouer
-        // `scale_frame` + `fit_dst_to_aspect` par-dessus appliquerait le padding deux fois et
-        // re-contiendrait une boîte déjà au bon ratio. Même raisonnement que pour la webcam.
-        let fit_screen = |dst: [f32; 4]| {
-            if app_screen_rect.is_some() {
-                dst
-            } else {
-                fit_dst_to_aspect(scale_frame(dst, padding_scale), crop_aspect)
-            }
-        };
-        // Issue #179 : le zoom se jouait entièrement dans la coupe source, donc la boîte
-        // écran restait au rect paddé et le zoom butait sur cette frontière au lieu
-        // d'atteindre les bords du cadre. On rend le zoom à la BOÎTE (cf. `remap_box`) :
-        // la coupe dessinée redevient le crop nu, la boîte porte le grossissement et
-        // déborde le padding — c'est la géométrie de `applyZoomTransform` (TS).
-        let s_base = fit_screen(p.screen.dst);
-        let s_base_prev = fit_screen(pp.screen.dst);
-        // Layouts "bloc" (side-by-side / top-bottom) : la boîte écran est un SLOT au ratio
-        // arbitraire, et le web y fait tenir l'image en `cover` (`computeCompositeLayout`
-        // renvoie `screenCover: true`, honoré par `frameRenderer`). Le natif l'ignorait, donc
-        // il étirait la source pour remplir le slot — visible dès que le clip est recadré,
-        // puisque le crop éloigne encore le ratio de la source de celui du slot.
-        //
-        // Le cover s'applique APRÈS le crop et le zoom, sur leur rect résultant : le crop
-        // décide quoi montrer, le zoom où regarder, le cover comment habiller la boîte. Son
-        // ratio de boîte se lit sur `s_base` : `remap_box` met les deux axes à la même
-        // échelle, donc la boîte finale a le même ratio et le cover ne dépend pas d'elle
-        // (ce qui casserait la circularité coupe → boîte → coupe).
-        let cover_box_ar = scene_ref.as_ref().and_then(|s| {
-            s.layout
-                .screen_cover
-                .then_some((s_base[2] * self.rw()) / (s_base[3] * self.rh()).max(0.0001))
-        });
-        let cover = |uv: [f32; 4]| -> [f32; 4] {
-            match cover_box_ar {
-                Some(ar) => cover_uv_rect(uv, [stw as f32, sth as f32], ar),
-                None => uv,
-            }
-        };
-        // La coupe RÉFÉRENCE (zoom entier) est celle qui remplissait la boîte paddée avant
-        // ce correctif ; la coupe DESSINÉE ne porte plus que le crop. `remap_box` reporte la
-        // seconde à travers le mapping de la première, ce qui conserve le cadrage exact.
-        // Le focus courant reste volontairement utilisé pour la frame précédente, comme avant.
-        let cut_ref = cover(screen_source_rect(u_max, v_max, active_crop, p.zoom, p.focus));
-        let cut_ref_prev = cover(screen_source_rect(u_max, v_max, active_crop, pp.zoom, p.focus));
-        let cut = cover(screen_source_rect(u_max, v_max, active_crop, 1.0, p.focus));
-        let s_dst = remap_box(s_base, cut_ref, cut);
-        let s_dst_prev = remap_box(s_base_prev, cut_ref_prev, cut);
-        // le padding n'affecte QUE l'écran (la quantité de fond révélée). La webcam reste ancrée
-        // en bas-droite à sa marge fixe, quelle que soit la valeur de padding (pas de scale_frame)
-        // — SAUF quand l'app a résolu un placement explicite (`app_webcam_rect`, drag-to-reposition
-        // compris). Ce rect est déjà exprimé en fraction du VRAI output (calculé côté web par
-        // `computeCompositeLayout` avec les vraies dimensions de sortie), position ET aspect déjà
-        // corrects — `fit_cam_aspect`/`scale_corner_br` (chemin preset par défaut) sont donc
-        // doublement inadaptés ici : ils réancrent au coin bas-droite (ignorant la position
-        // choisie par l'utilisateur) ET recalculent l'aspect en pixels du canvas fixe 16:9
-        // (`OUT_W`×`OUT_H`), une référence différente du vrai output dès que la sortie n'est pas
-        // 16:9 (rapport utilisateur : webcam glissée au coin bas-gauche en 9:16, JSON envoyé au
-        // natif confirmant une position flush, mais rendu native visiblement décalé ET trop
-        // petit). On garde seulement `scale_center` (zoom réactif, préserve position+aspect) puis
-        // on pré-compense par `inverse_undistort` pour annuler le `undistort()` générique
-        // appliqué plus bas à tous les calques (écran compris) — sans quoi ce rect déjà correct
-        // se ferait déformer une seconde fois par cet undistort partagé.
-        let mut w_dst = if app_webcam_rect.is_some() {
-            scale_center(p.webcam.dst, webcam_size_scale)
-        } else {
-            fit_cam_aspect(scale_corner_br(p.webcam.dst, webcam_size_scale))
-        };
-        let mut w_dst_prev = if app_webcam_rect.is_some() {
-            scale_center(pp.webcam.dst, webcam_size_scale_prev)
-        } else {
-            fit_cam_aspect(scale_corner_br(pp.webcam.dst, webcam_size_scale_prev))
-        };
 
-        // Full Camera : la caméra PREND le cadre — parité `computeCameraFullscreenRect` (TS).
-        // La cible est exactement [0,0,1,1] : pas de marge, pas de padding, pas d'arrondi, et
-        // plus rien de la composition (fond, écran, ombre) derrière. Le rect change de ratio en
-        // chemin, mais `cover_crop_uv` (plus bas) dérive la coupe source du ratio RÉEL de la
-        // boîte à chaque frame : la caméra n'est donc jamais étirée pendant l'animation.
-        let fullscreen_dst = |dst: [f32; 4], progress: f32| -> [f32; 4] {
-            if progress <= 0.0 {
-                return dst;
-            }
-            let lerp = |a: f32, b: f32| a + (b - a) * progress;
-            [lerp(dst[0], 0.0), lerp(dst[1], 0.0), lerp(dst[2], 1.0), lerp(dst[3], 1.0)]
-        };
-        // Petit côté de la boîte caméra AVANT que Full Camera ne la fasse grandir. C'est la
-        // référence du rayon de coin : le zoom réactif est déjà dedans (il rétrécit la boîte,
-        // donc l'arrondi suit tout seul — parité `borderRadius * reactiveFactor` côté TS), alors
-        // que Full Camera ne fait pas grossir l'arrondi, il le DISSOUT (cf. `shape_fade`).
-        let w_nominal_min = (w_dst[2] * self.rw()).min(w_dst[3] * self.rh());
-        w_dst = fullscreen_dst(w_dst, cam_progress);
-        w_dst_prev = fullscreen_dst(w_dst_prev, cam_progress_prev);
-
-        // Contre-étirement "fit" : le canvas interne compose TOUJOURS en OUT_W×OUT_H (16:9),
-        // puis `blit_resized` étire tout, de façon non uniforme si besoin, vers la résolution
-        // de sortie demandée — voulu pour que le FOND (dessiné plus bas en dst=[0,0,1,1])
-        // remplisse tout le cadre quel que soit le ratio choisi. Mais l'écran et la webcam ne
-        // doivent PAS être déformés par cet étirement : on rétrécit ici leur rect de
-        // destination (centré, dans cet espace 16:9 PRÉ-étirement) par l'inverse du plus fort
-        // des deux facteurs d'étirement, pour qu'après l'étirement final leur ratio d'origine
-        // reste préservé (letterboxé/pillarboxé sur le fond, qui lui reste plein cadre) — mode
-        // "fit"/contain. Si l'utilisateur veut un rendu "fill" (remplir sans bandes), il ajuste
-        // le crop lui-même ; le natif ne fait plus ce choix à sa place en étirant l'image.
-        // Le dessin du coin (SDF, shaders.hlsl) compare le rayon à `quad_px`, exprimé en px du
-        // RENDER TARGET : c'est donc dans cet espace-là qu'il faut le lui donner.
-        //
-        // Toutes les longueurs de la scène sont des FRACTIONS ; on les multiplie ici par ce
-        // qu'elles mesurent, dans l'espace du render target. C'est ce qui rend preview et export
-        // identiques : « un pixel » n'y désigne pas la même chose (la preview rastérise dans un
-        // cadre contain-fitté plus petit, cf. `preview_render_size`), alors qu'une fraction, si.
-        // `frame_min_px` est la référence des quantités relatives au CADRE ; un rayon de coin,
-        // lui, se mesure contre sa propre boîte — il doit rester en place quand on redimensionne
-        // la boîte, pas suivre le cadre.
-        let frame_min_px = self.rw().min(self.rh());
-        let s_min_px = (s_dst[2] * self.rw()).min(s_dst[3] * self.rh());
-        let app_screen_radius_frac = scene_ref.as_ref().and_then(|s| s.layout.screen_radius_frac);
-        let scene_roundness_frac = scene_ref.as_ref().map(|s| s.effects.roundness_frac);
-        // Le rayon suit la boîte : quand le zoom l'agrandit (issue #179), les coins grandissent
-        // avec elle puis sortent du cadre — comme le masque de la référence, qui porte le même
-        // `br: maskBorderRadius * camS` et quitte l'étage au même moment.
-        let s_radius = match (cfg.rounded, app_screen_radius_frac, scene_roundness_frac) {
-            (false, _, _) => 0.0,
-            // Preset en bloc : le rayon appartient à la boîte écran (parité exacte avec la caméra).
-            (true, Some(f), _) => f * s_min_px,
-            // Scène sans rayon imposé : slider Roundness, relatif au cadre.
-            (true, None, Some(f)) => f * frame_min_px,
-            // Fixture/bench (pas de scène) : chemin inspector historique, inchangé.
-            (true, None, None) => p.screen.radius * lp.radius_scale,
-        };
-        let w_px = [w_dst[2] * self.rw(), w_dst[3] * self.rh()];
-        // Rayon caméra. Le slider Roundness ne s'y applique jamais (il ne vaut que pour l'ÉCRAN).
-        // Quand l'app le résout (`computeCompositeLayout`, source unique), on le prend : c'est la
-        // seule façon que les deux moitiés d'un layout en bloc soient encadrées à l'identique,
-        // l'écran consommant déjà `screen_radius_frac` du même calcul. La table ci-dessous en
-        // était une SECONDE, indépendante — fraction différente (0.12 vs 0.06 côté web) et sans
-        // bornes — donc écran et caméra ne pouvaient pas s'accorder.
-        let app_webcam_radius_frac = scene_ref.as_ref().and_then(|s| s.layout.webcam_radius_frac);
-        // Full Camera dissout la forme en même temps qu'elle prend le cadre : le rayon fond
-        // vers 0 avec `cam_progress`, donc le cercle devient un rect à coins de plus en plus
-        // francs puis un plein cadre net — aucun masque ne survit au plein écran (parité
-        // `computeCameraFullscreenRect`, qui ramène `maskShape` à "rectangle" et lerpe le
-        // rayon vers 0 pour exactement la même raison).
-        let shape_fade = (1.0 - cam_progress).clamp(0.0, 1.0);
-        let w_radius = shape_fade
-            * w_nominal_min
-            * match app_webcam_radius_frac {
-                Some(f) => f,
-                // Fallback (payload sans fraction, fixture/bench) : l'ancienne table, keyée sur la
-                // forme. Rectangle ET square n'ont qu'un léger arrondi (0.12) et ne diffèrent que
-                // par le ratio ; rounded est nettement plus arrondi (0.3) ; circle = demi-côté.
-                None => match lp.webcam_shape {
-                    1 => 0.5,
-                    3 => 0.3,
-                    _ => 0.12,
-                },
-            };
 
         self.begin([0.0, 0.0, 0.0, 1.0]);
